In DSP_GetRegisterAddress, replace size_t with int

+++ src/falcon/dsp.c
@@ -593,7 +593,7 @@ int DSP_GetRegisterAddress(const char *regname, uint32_t **addr, uint32_t *mask)
 	typedef struct {
 		const char name[MAX_REGNAME_LEN];
 		uint32_t *addr;
-		size_t bits;
+		int bits;
 		uint32_t mask;
 	} reg_addr_t;
